change log info location from .h to .c
